Throw compile error message on program build
Previously when a compile error occurred on
some kernel source code, it would simply throw
an empty exception with no indication as to
what had occurred during the build of the kernel.

Instead, we pull the build log for each device
available to the GPU kernel compiler and return
it in a new exception so the user is notified.